+++ js/packages/openinference-instrumentation-langchain/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @arizeai/openinference-instrumentation-langchain
 
+## 4.0.16
+
+### Patch Changes
+
+- Updated dependencies [0168198]
+  - @arizeai/openinference-core@2.5.0
+
+## 4.0.15
+
+### Patch Changes
+
+- Updated dependencies [145e3c6]
+  - @arizeai/openinference-semantic-conventions@2.6.0
+  - @arizeai/openinference-core@2.4.1
+
 ## 4.0.14
 
 ### Patch Changes