+++ js/packages/openinference-instrumentation-mcp/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @arizeai/openinference-instrumentation-mcp
 
+## 0.2.22
+
+### Patch Changes
+
+- Updated dependencies [0168198]
+  - @arizeai/openinference-core@2.5.0
+
+## 0.2.21
+
+### Patch Changes
+
+- Updated dependencies [145e3c6]
+  - @arizeai/openinference-semantic-conventions@2.6.0
+  - @arizeai/openinference-core@2.4.1
+
 ## 0.2.20
 
 ### Patch Changes